+++ MovieQuiz/Models/QuizQuestion.swift
@@ -1,0 +1,14 @@
+//
+//  QuizQuestion.swift
+//  MovieQuiz
+//
+//  Created by Анастасия  Колгушкина  on 09.03.2025.
+//
+
+import Foundation
+
+struct QuizQuestion {
+    let image: String
+    let text: String
+    let correctAnswer: Bool
+}
