+++ MovieQuiz/Models/QuizResultsViewModel.swift
@@ -1,0 +1,8 @@
+//
+//  QuizResultsViewModel.swift
+//  MovieQuiz
+//
+//  Created by Анастасия  Колгушкина  on 09.03.2025.
+//
+
+import Foundation
