+++ MovieQuiz/Presentation/MovieQuizViewController.swift
@@ -2,94 +2,40 @@ import UIKit
 
 final class MovieQuizViewController: UIViewController {
 
-    struct QuizQuestion {
-        let image: String
-        let text: String
-        let correctAnswer: Bool
-    }
-
-    private let questions: [QuizQuestion] = [
-        QuizQuestion (
-            image: "The Godfather",
-            text: "Рейтинг этого фильма больше чем 6?",
-            correctAnswer: true),
-        
-        QuizQuestion (
-            image: "The Dark Knight",
-            text: "Рейтинг этого фильма больше чем 6?",
-            correctAnswer: true),
-        
-        QuizQuestion (
-            image: "Kill Bill",
-            text: "Рейтинг этого фильма больше чем 6?",
-            correctAnswer: true),
-        
-        QuizQuestion (
-            image: "The Avengers",
-            text: "Рейтинг этого фильма больше чем 6?",
-            correctAnswer: true),
-        
-        QuizQuestion (
-            image: "Deadpool",
-            text: "Рейтинг этого фильма больше чем 6?",
-            correctAnswer: true),
-        
-        QuizQuestion (
-            image: "The Green Knight",
-            text: "Рейтинг этого фильма больше чем 6?",
-            correctAnswer: true),
-        
-        QuizQuestion (
-            image: "Old",
-            text: "Рейтинг этого фильма больше чем 6?",
-            correctAnswer: false),
-        
-        QuizQuestion (
-            image: "The Ice Age Adventures of Buck Wild",
-            text: "Рейтинг этого фильма больше чем 6?",
-            correctAnswer: false),
-        
-        QuizQuestion (
-            image: "Tesla",
-            text: "Рейтинг этого фильма больше чем 6?",
-            correctAnswer: false),
-        
-        QuizQuestion (
-            image: "Vivarium",
-            text: "Рейтинг этого фильма больше чем 6?",
-            correctAnswer: false)
-    ]
-    struct QuizStepViewModel {
-      let image: UIImage
-      let question: String
-      let questionNumber: String
-    }
                     
     private var currentQuestionIndex = 0
-    
     private var correctAnswer = 0
-    
     private var isShowingResult = false
     
     @IBOutlet private var imageView: UIImageView!
-
     @IBOutlet private var textLabel: UILabel!
-    
     @IBOutlet private var counterLabel: UILabel!
     
+    private let questionsAmount: Int = 10
+    private var questionFactory: QuestionFactory = QuestionFactory()
+    private var currentQuestion: QuizQuestion?
     
     
     override func viewDidLoad() {
         super.viewDidLoad()
-        let currentQuestion = questions[currentQuestionIndex]
-        let viewModel = convert(model: currentQuestion)
-        show (quiz: viewModel)
+        
+        if let firstQuestion = questionFactory.requestNextQuestion() {
+            currentQuestion = firstQuestion
+            let viewModel = convert(model: firstQuestion)
+            show(quiz: viewModel)
+        }
+       // let firstQuestion = questions[currentQuestionIndex]
+       // let viewModel = convert(model: currentQuestion)
+        //show (quiz: viewModel)
     }
     
     @IBAction private func yesButtonClicked(_ sender: Any) {
         if (!isShowingResult) {
+            guard let currentQuestion = currentQuestion else {
+                return
+            }
             
-            if questions[currentQuestionIndex].correctAnswer {
+            if currentQuestion.correctAnswer {
                 showAnswerResult(isCorrect: true)
             }
             else {
@@ -99,7 +45,14 @@ final class MovieQuizViewController: UIViewController {
     }
     @IBAction private func noButtonClicked(_ sender: Any) {
         if (!isShowingResult) {
-            if !questions[currentQuestionIndex].correctAnswer {
+            // было
+           // let currentQuestion = questions[currentQuestionIndex]
+
+            // стало
+            guard let currentQuestion = currentQuestion else {
+                return
+            }
+            if !currentQuestion.correctAnswer {
                 showAnswerResult(isCorrect: true)
             }
             else {
@@ -112,7 +65,7 @@ final class MovieQuizViewController: UIViewController {
         let questionStep = QuizStepViewModel(
             image: UIImage(named: model.image) ?? UIImage(),
             question: model.text,
-            questionNumber: "\(currentQuestionIndex + 1)/\(questions.count)")
+            questionNumber: "\(currentQuestionIndex + 1)/\(questionsAmount)")
         return questionStep
     }
     
@@ -140,30 +93,43 @@ final class MovieQuizViewController: UIViewController {
     private func showNextQuestionOrResults() {
         imageView.layer.borderWidth = 0
         isShowingResult = false
-      if currentQuestionIndex == questions.count - 1 {
-          let alert = UIAlertController(
-            title: "Этот раунд окончен!",
-            message: "Ваш результат - \(correctAnswer) из \(questions.count)",
-            preferredStyle: .alert)
-          
-          let action = UIAlertAction(title: "Сыграть еще раз", style: .default) {
-[weak self] _ in
-              guard let self = self else { return }
-              self.correctAnswer = 0
-              self.currentQuestionIndex = 0
-              let currentQuestion = self.questions[self.currentQuestionIndex]
-              let viewModel = self.convert(model: currentQuestion)
-              self.show (quiz: viewModel)
-          }
-          
-          alert.addAction(action)
-          self.present(alert, animated: true, completion: nil)
-      } else {        
+        if currentQuestionIndex == questionsAmount - 1 {
+            let alert = UIAlertController(
+                title: "Этот раунд окончен!",
+                message: "Ваш результат - \(correctAnswer) из \(questionsAmount)",
+                preferredStyle: .alert)
+            
+            let action = UIAlertAction(title: "Сыграть еще раз", style: .default) {
+                [weak self] _ in
+                guard let self = self else { return }
+                self.correctAnswer = 0
+                self.currentQuestionIndex = 0
+                //  let currentQuestion = self.questions[self.currentQuestionIndex]
+                //  let viewModel = self.convert(model: currentQuestion)
+                // self.show (quiz: viewModel)
+                
+                if let nextQuestion = questionFactory.requestNextQuestion() {
+                    currentQuestion = nextQuestion
+                    let viewModel = convert(model: nextQuestion)
+                    show(quiz: viewModel)
+                }
+            }
+            alert.addAction(action)
+            self.present(alert, animated: true, completion: nil)
+        }
+        
+        else {
         currentQuestionIndex += 1
-        let nextQuistion = questions[currentQuestionIndex]
-        let viewModel = convert(model: nextQuistion)
+      //  let nextQuistion = questions[currentQuestionIndex]
+      //  let viewModel = convert(model: nextQuistion)
           
-        show(quiz: viewModel)
+       // show(quiz: viewModel)
+          
+            if let nextQuistion = questionFactory.requestNextQuestion() {
+                currentQuestion = nextQuistion
+                let viewModel = convert(model: nextQuistion)
+                show(quiz: viewModel)
+            }
       }
     }
 }
